Removing the reliance on __COUNTER__
Instead there is a global volatile counter variable in the
implementation of the profiler which get incremented via
_InterlockedIncrement. The macros have been adjusted to use a static
variable to each function block that is called multiple times still uses
the same index. This has been done because __COUNTER__ depends on the
current translation unit and is unique for each translation unit.
Instead the entry index should be unique on a binary level instead.

Despite the use of an interlocked increment on a volatile global this
code is not thread safe. Or rather it has not been tested in
multithreading scenarios. The reason is that for the time being I do not
have the need for such capabilities. Once the need arises I will
probably modify this library and make it thread-safe in some shape or
form.