Showing inclusive and exclusive timings in the profiling output
A profiling zone shows its total runtime that also includes child zone
timings. This is called the inclusive time. However, sometimes it is
good to know how much time is spent in the function without including
all the sub profiling zones as well. So it should subtract the total
runtime of all sub zones from the total runtime of the current profiling
zone. This is called the exclusive time (it exclused sub zones).

This is adequate but also breaks when profiling recursive functions.