Breaking change to dani_Profile, dani_ProfileBandwidth, dani_ProfileFunction, and dani_ProfileFunctionBandwidth macros
The functions named in the title used to take in a statement block of C
code. This had the benefit of not having to worry about callind end by
hand. However, after using these macros for a while there have been some
downsides to it that I didn't like.

1) Error reporting of the MSVC compiler becomes basically useless. If
   you change something in the code you want to profile and accidentally
   make an error, the compiler will no longer report the correct line in
   the source file but rather point to the beginning of the macro
   invocation. This lead to having to comment out the profiling macro
   when making changes to the code, only to then uncomment it again
   after I verified everything was working correctly. However, I also
   sometimes forgot to uncomment the profiling macro again which lead to
   a whole run of a benchmark without any profiling enabled... annoying.

2) Profiling sub zones became a mess fast, especially when trying to
   access a variable declared within a different profiling zone. I had
   to arkwardly reformat my code so I could profile the necessary blocks
   I wanted. And I do not like having to format my production code only
   to make profiling work. Profiling a section of code should be as
   unintrusive as possible. These old style macros were not!

Not a lot of downsides but massive enough to make me change how they
work. I rather deal with accidentally forgetting a dani_ProfileEnd as
opposed to the things mentioned above. (Defer would help a lot here but
C does not have it...)